+++ lib/flutter_image_cropper.dart
@@ -1,6 +1,7 @@
+
 import 'flutter_image_cropper_platform_interface.dart';
 
-export 'flutter_image_cropper_platform_interface.dart' show CropAspectRatioPreset;
+export 'flutter_image_cropper_platform_interface.dart' show CropAspectRatioPreset, CropShape;
 
 class FlutterImageCropper {
   static Future<String?> cropImage({
@@ -9,6 +10,7 @@ class FlutterImageCropper {
     double? aspectRatioY,
     List<CropAspectRatioPreset>? aspectRatioPresets,
     int quality = 90,
+    CropShape cropShape = CropShape.rectangle,
   }) {
     return FlutterImageCropperPlatform.instance.cropImage(
       sourcePath: sourcePath,
@@ -16,6 +18,7 @@ class FlutterImageCropper {
       aspectRatioY: aspectRatioY,
       aspectRatioPresets: aspectRatioPresets,
       quality: quality,
+      cropShape: cropShape,
     );
   }
 }